fea(agent/test-framework) new agent integration test framework
This new agent integration test is heavily based on AbstractAgentTest and AgentTestContext. The structure has been re-made
a bit so that the provider configuration is also tested, instead of specifying an agent directly.

It integrates with the agent-test-server by saving/loading credentials and asking supplemental information.

Currently it does minimal data quality/sanity checks, it's a future todo. It "pretty" prints almost all collected data
in an account hierarchy (meaning account is the top object).