fix(refreshableItem) add legacy items to sort order
The previous solution tried to remove the legacy items from the list to be sorted.
Since you cannot sort a list containing items that are not included in the sort order.

This commit stops deleting items (which was a problem when immutable lists where used) and instead adds the legacy items to the sort order.