fea ImageRecognizer
This is a move / re-implementation of https://github.com/tink-ab/tink-backend/pull/7212

Everything except the changes to the "Dockerfile" is done. In addition all 3:d party libraries were updated to their latest versions.

The previously existing unit test works.

I have sucessfully used this to interpret images received from ING ES (Spain), but it turned out recognising the images is not necessary. So I will not us this now...