fea(Revolut) transform entity to Tink account
Revolut has a "wallet" of "pockets", where each pocket has a separate currency, and can have either one or two accounts associated with it. It's the pockets that are transformed into a Tink accounts, so when a pocket has multiple associated accounts we pick the one that is displayed in the Revolut app when requesting details required for inserting money via bank transfer.