fea(Nordea-fi) Implemented new Nordea-fi agents
Nordea-fi released a new app that makes use of a completely new API.
This commit implements a new agent that uses this API.

**Notes:**

- Old Nordea agent has been moved into package `nordea.v21`, while the new
agent is implemented in package `nordea.30`

- The new agent only allows sign in using Nordea Codes app which is
similar to bankid. This is achieved through implementing the
`ThirdPartyAppAuthenticator`.

- The new api seems to have some bugs. Some of the fields in the JSON
response model are always null, eg. `market` in  `InstrumentEntity`.
For now this is solved by getting what is needed from other fields
where possible.

- Since our credentials does not have any loans NordeaFiLoanFetcher
logs the responses for now. There is an implementation of loan accounts
based on test files found when unpacking the Nordea app, but these are
not tested.

- The code for automatically opening Nordea Codes when authenticating
is not tested for android.

- Seeding for previous version of Nordea app has been removed.

- Updated old version of nordea.v21 to use unique mandatory uniqueId
since it generates build errors otherwise.