fix(Revolut) Fixed errors in authentication
Two factor authentication requires, aside from phone number, a user specific four digit passcode, in order to be initiated. In the previous implementation this code was believed to be device specific, and thus generated automatically, which lead to a 401 response from the API.

The default User-Agent header causes a 400 response, so the one used by the Revolut app was modified to include Tink's value.