fix(abnamro) Remove unsupported update action
Ensure that the user being updated isn't set to "UPDATED"
pre-emptively to prevent a race condition.

For ABN AMRO aggregation is only used to mark which users are active and
then transactions are sent via the connector. By calling
context.updateAccount directly we created a race condition where
aggregation might pre-emptively set an account as done which would
prevent the connector from sending transactions to that user.