(Fix) Changed the NextGen user agent to use the new message.
If there is no AgentContext available (due to testing), it is set to "unknown aggregator"
However, the user agent should be changed to the x-aggregator?
'

(Fix) Added the new X-Aggregator header to the TinkHttpClient

(Fix) Added header to the message sent as an X-Aggregator header.

The X-Aggregator value is based on the passed context value, where an aggregator is set.
If there is no context available, the value is set to Unknown Aggregator