ref(update-account): split method into two methods
First method caches all accounts. It is still called updateAccount
and is called from all Agents)

Second method is called sendAccountToUpdateService. It takes an account
from the cache and sends to UpdateService.

I also changed updateAccount to not return an Account anymore.

NOTE: This commit does not work on it's own and requires the next commit
to work. This split was done for reviewer convenience.