fix(uk/revolut) transactions stuck fetching
The Revolut API lets us fetch transactions up until a given timestamp (ms), which can be extracted from the transaction entities. Instead of using these given timestamps, wrongly converted the Date stored in the new Transaction objects, which due to rounding of time (to 12 am) resulted in the transaction fetcher being stuck in an infinite loop in some cases.