fix(uk/revolut) mixed transaction currencies
Revolut doesn't use accounts in the manner we are used to, but splits transactions/payments/etc by currency. This was not echoed when fetching transactions, so the resulting list contained all transactions together, with mixed currencies. Fixed this by adding currency to the temporary storage of the account, and checking that the transaction and account currencies match.