fea(serializing) annotation for double values
Sometimes requests have doubles in them. This is unsafe for inaccuracy reasons, where the serialized values may get a long train of decimals in them. This annotation will by default output the serialized double as a number with two decimals (ROUND_HALF_UP), but also gives the option to set the number of decimals, turn on/off trailing zeros, output as String instead of numerical value etc.