fea(AutomaticRefresh) Persisting jobs util they either fail or are done
The jobs where previously deleted directly from the queue without confirming that the job either failed or was done. To create a more reliable service the job is now consumed from the queue and hidden from other consumers for 5 minutes before it is visible again. If the job is completed (with or without failures) it will be deleted from the queue. Further on this should be changed to give the option to not remove the job for temporary failures.