fix(AMEXv45) no transactions for cards
There were few problems. First of them was true/false values switched when
checking if we get error response from bank. Then when these were fixed,
we did not take info account that there may be no transactions for last
month, and we got NullPointer as message instead of list of transactions
was present.
Biggest problem was assumption that identifiers for cards are constant.
If we get few cards in response the cards will get consequtive numbers,
but then if we request for details for one card, it will be the only
card on this page, so the index will be `0`, even if before it was `3`.
Because of that, when we done the check if a transaction was performed
on a given account we filtered out all of them, as the indexes did not
correspond to each other.