fix(Revolut) pocket with no associated accounts
Note: For Revolut we transform "Pockets" to Tink accounts. A pocket is an abstraction of a currency account, and may have zero, one or two underlying bank accounts.

We could previously only handle cases of pockets with one or two associated accounts, and any other case would result in a `IllegalStateException` being thrown. Now pockets with no accounts are filtered out, as we have no way of setting an appropriate `uniqueIdentifier` for Tink accounts based on such pockets, which would lead to duplicate Tink accounts being created if one of those pockets were to ever get an associated account.