fix(AutomaticRefresh) Queue logic for auto-refreshes
The consumer will take an automatic refresh and hide it (prevent it) from being consumed in 10 minutes. During these 10 minutes this item will be queued in the provider specific in-memory queue of a size of 10. However, if the provider specific job is full a rejectedException will be thrown and catched. This job will not be executed. However, it will not be deleted from the SQS queue, but stored for later consumation (10 minutes). If the job finishes (with or without exception) it will be deleted as usual. The rejectedException is passed via the autmaticRefreshStatus.