fix(AutomaticRefresh) Removed unused functions and implement serial exec
Instead of keeping track of the threads and their executing stages the addition of new jobs from the SQS queue to the message queue is now serial. This is more reliable and increases readability. However, it will have a negative impact on processing throughput, due to serialization. In addition jobs can be lost if the service goes down. This is due to jobs being deleted before they are commited. To reduce the amount of jobs potentially lost, the inmemory queue is decreased to only contain one job.