tes(signature) Added signatureKeyPair to tests
Previously `X-Signature` and `X-Signature-Info` were only added to
http headers in production, these are now also added to headers in
development so that they appear in tests.

This was done by adding signatureKeyPair configurations in
development.yml, which is used to build the ServiceConfiguration for the
agents when running tests.