fix(danskebank) Trimmed header to acceptable size
Removed `x-app-culture` from DanskeBankFI header. This was done to bring
the number of header fields to 17, which is for some unkown reason
the maximum that the DanskeBank server will handle. This was only done
for DanskeBankFI since it alone had enough headers for there to be a
problem.

If this limit is not taken into account the request will generate a
`400 Bad Request` response with the following text.

```
Your browser sent a request that this server could not understand.
The number of request header fields exceeds this server's limit.
```

This error seems to have been introduced in commit [90f195a7eb8187175261cf8e7f5258b593e83e5e](https://github.com/tink-ab/tink-backend-aggregation/commit/90f195a7eb8187175261cf8e7f5258b593e83e5e#diff-4caf6a91be5bbcaf32abd71f94fe8dfd)
when we started sending `X-Signature` in the header, since this brought
the header over the limit.