fix(amexV3) do not send profileInfo upon login to fix auth errors
This "fix" is a test fix, not sure if it will work or if it's the cause for the errors.

This commit removes sending the profileInfo field (which we get after a successful login) when we do a login. This information originates from the old aggregation cluster and could be the reason for why it doesn't work in the new cluster (new IP).