fea(AutomaticRefresh) If message is rejected, put it back in queue
If a message is rejected, it will be requeued. To avoid unfavourable patterns in the queueue, the requeueing mechanism will queue the items with a random delay between 0 - 900 seconds  (0-15 minutes). This fix will help the service to acheive a higher throughput by filling other ratelimited queues, instead of waiting for one queue to become free.

NOTICE:
Re-queueing items will increase the number of produces to the queue drastically and thus, should be fixed in a later PR.