fix(se/shb) get legacy values from sensitivePayload
Old users will have their pinned (libtfa) data on the sensitivePayload directly and not in the persistent storage. This, somewhat temporary, fix tries to get
the data from persistent storage and if that fails from the sensitivePayload.

The data will always be stored on persistentStorage, and removed from the sensitivePayload, when persisted.