fea(es/bankia) output creditCard transactions
The previous code logged credit card transaction information wrt pagination. So far we haven't received any good logs.

This commit changes so that the transactions are outputted. And if the bank indicates it has more we log it.

If we figure out how to paginate we must remember to fetch ALL transactions regardless of certain date the first time.