fea(url-query) Ability to append full url query
Added the ability to append a fully formated query to a http request.

This is usefull in cases where a full query is provided by the host and
it just needs to be added to a subsequent request.

This method performs no checks or sanitisation in the way
`URL.queryParam`.