fea(labanquepostale-auth) LaBanquePostale Authentication
Implemented Authenticatior and SessionHandler for LaBanquePostale.

**Notes**

The Authenticatior makes use of image recognition to parse and a jumbled
numpad which the user would normally enter their passcode on.

The banks API relies on redirection, meaning that it will either send
the requested data to the client or redirect it to an error page.
The status of responses is therefore extracted from the final
redirection URL for each requst in the `LaBanquePostaleApiClient`.