fix(OP) Use marketValue if marketPrice is unavailable
For some users the priceEur field is empty for marketValue, this causes a NPE when we try to fetch the instrument price. Since marketValue seems to be available in cases when marketPrice isn't we're using that as a backup, calculating the instrument price as marketValue divided by the quantity. And if that is also null we don't set the instrument price.