fix(loan-type) Moved LoanInterpreter from backend to backend-aggregation
This commit fixes the issue where loan types that were previously
correctly set would suddenly become of type `OTHER`. This commit fixes
the issue without any changes to `backend`, although some features
can now be removed from `backend`.

**Bug Cause**

Code for setting loan type was split up between `backend` and
`backend-aggregation`. This caused a problem when `backend-aggregation`
stopped sending `null` for unkown loan types. The `null` loan type would
trigger `backend` to guess the loan type from the name of the loan, but
 since `backend-aggregation` would now send `OTHER` this never happened.

**Solution Summary**

Moved `LoanInterpreter` to `backend-aggregation`. The loan type sent to
`backend` is now final. The loan type goes through the following steps
before it is sent to `backend`.

1. Explicitly set to a known type or the new `DERIVE_FROM_NAME` type
in the agent.

2. If `type == DERIVE_FROM_NAME` try to guess type based on name.

3. If the type is not explicitly set and coult not be guessed type
is set to `OTHER`.

**Implementation**

- The `LoanDetails.setType()` has been removed. Instead a type must be
given in the builder constructor. In cases where the type is not known
`Type.DERIVE_FROM_NAME` can be given.

- When the loan in converted to a system type before it is sent to
`backend` the type is checked, if of type `DERIVE_FROM_NAME` a
LoanInterpreter based on the agents market will be used to guess the type.

- It is impossible to send `null` loan type to `backend` if one is using
a NextGenerationAgent or the LoanDetails builder model. But a
Precondition in LoanDetails.Type still exists to make sure this does
not happen in case of code changes in the future.

- Agents that makes use of LoanDetails has been updated to take these
changes into account.