fix(SparbankenSyd) Fetch credit card and transactions for SparbankenSyd
This code contains multiple ways of fetching data from backend.
We have one user at the moment with a credit card and are trying to
find the correct way to fetch credit cards for Sparbanekn syd.
Once we have a working flow, we can clean up the code. If it is very
different from DK/NO SE can have their own CreditCard fetcher