fet(euroInf) Add support for transaction pagination
Not all banks use pagination, only the ones using PMF in their backend.
Every bank is able to return transactions for last month using old, not
paginated endpoint, but not other way around.
This solutioni creates fetchr for paginated  and not paginated
transaction history, using not paginated as default. Pagination setting
can be set in the Configuration of given bank.