fix(nxgen-base-test) Fixed crash caused by MarkedCode
**Bug Description*

The partial introduction of MarketCode caused a crash when
NextGenerationAgentBaseTest set `market=TEST` since this market
was not suppored by the enum.

**Solution**

This commit introduces `TEST` market as a temporaty fix,
since there is an open ticket for properly integrating MarketCode.

The code that converts `String market` to `MarketCode market`
has also been moved and commented to be more obvious.