fea(handelsbanken-se-fi) Added transaction paginator
- Implemented transaction account paginator for handelsbanken-fi and
handelsbanken-se.

- Refactored old code such that api calls are not made from entities
when fetching transactions.

- Added `HandelsbankenSEUpcomingTransactionFetcher`. Fetching upcoming
transactions was previously done in `TransactionsSEResponse`.

- Added method `getURL()` to `URL` which will return a new `URL` object
with the same url value but without the query.

**Pagination Implementation**

For handelsbanken-fi the bank API provides an url to
"more transactions" for each call, so a standard key paginator is used.

For handelsbanken-se the bank API provides an url that will list
transactions between two indices, but the default values for the indices
are `from: -1, to: -1` which will cause the API to list what
appears to be the Last three months. Here an index paginator is used and
the default indices are replaced with the pagination indices.