fea(euroInformation) Enabling PMF if possible
Backend support 2 types of requests, one when PMF is enabled and one
when it's not. With PMF enabled we can get the whole history and some
requests are build a bit differently. This commit adds enabling PMF, when
possible, after authentication and storing info in sessionStorage to
later make a decision which fetcher should be used.