fix(AgentDebug) Mask fields marked as masked in the provider
Previously only passwords where masked in the agent-debug. This commit reads the provider fields for a new debug and cross-references it to the fields in the credentials. If the fields are marked as masked in the provider and is present in the credential, the String.replace will replace the value of the field present in both the provider and credential.*

The method has been tested by connecting the phone to the local backend and verifying the "before and after" logs. In addition, it has been tested with both password verification, pin verification and with bankID (to verify that the 12 and 10 digit personal number is masked when marked as masked in the config files)

fix(AgentDebug) Sensor all fields present in the provider