fea(ClusterIdProvider) create cluster ID provider for provider service
I dont know if this is necessary or not, but I duplicated the `ClusterInfoProvider` for this service because this is using the ClusterHostConfiguration db repo. For Provider Service, and potentially other services that does not need the `base64encodedclientcertificate` and other fields in the db, this is not necessary. This provider only exposes the cluster id and remove the dependency on aggregation db.