fix(bawagpsk) Agent can now fetch accounts beyond process lifetime
If loadCredentialsBefore and saveCredentialsAfter
are true, and the agent integration test is run
twice, the agent would crash. This is because the
agent requires data from the login response, which
does not exist in the second run because
authentication is not done in that case. The data
needed from the login response is now saved to
peristent storage at authentication.