tes(agent) Added expectLoggedIn flag to agent test
Added expectLoggedIn that allows agent tests to opt-out of the
`Assert.assertTrue("Expected to be logged in.", keepAlive(agent));`
check in AgentIntegrationTest by setting this flag to `false`.

This allows agents to implement a SessionHandeler that always
throws `SessionError.SESSION_EXPIRED` if this is required by the
authentication flow and still pass the test.

The flag defaults to `true`.