fix(AgentDebug) Changed the logged message to return the AWS CLI URI
The current implementation of the S3 storage for debugs returns the HTTP URI where the file is located. However, this is not practical for the integration squad since they only have access to the files via the CLI. Thus this PR changes the displayed value to be the CLI URI instead of the HTTP one.