feature: only query provider status table once
The current solution queries the provider status table for every
provider, i.e. a long list of provider will query many times.
This is a big drawback on performance.

The new way is to only query the table once per request and create
a map of all providers which we use to lookup the provider status.