ref(ukob-ais) Refactored base agent construction
Refactored base agent such that implementations of this agent will
construct their own pagination controllers instead of just the paginator.

This is meant to allow for more flexibility in the future, in case the
type of pagination changes etc.

Also changed return type of `makeUpcomingTransactionFetcher` to `Optional`
in order to indicate that it is in fact optional.