fea(ukob-ais) Switched from bulk balance to per-account balance fetch
Previously `UkOpenBankingAccountFetcher` used "bulk fetching" for both
accounts and balances, and then mapped a balance to each account. Sadly
the bulk balance api service is considered "optional", so it is likely that
many banks won't implement it.

For this reason account balance fetching is now done by streaming through
all accounts, fetching the balance for each one.

A few interfaces was needed in order to make the mapping of balances to
accounts work. Mainly the ability to stream through all accounts and
get their `accountId`.