fix(ukob-ais) Signed transactions
Ukob transactions are now signed based on their `BankTransactionCode`.

The Ukob standard relies on long form `ISO 20022` codes in order to
determine the type of transaction. This code can also be used to
determine if the transaction is incoming or outgoing. This commit
adds an enum containing all of these codes, sorting these codes into
outgoing and incoming.

It also replaces the previous method of instantiating enums such that
enums can be upercase underscored format, which should improve
readability when dealing with names containing many words.