fix(se/swedbank) do not output duplicate transactions
Swedbank has a bug where they give us the same page multiple times during a pagination.

This commit solves it by storing the earliest date we have seen so far and compare it against the fetched page.
If we have seen the page before we return an empty list of transactions but with the correct next link so that we can fetch the next page.