fix(ukob-ais) Account balance is retrieved in a more sensible way
A Ukob account can include several instances of `BalanceEntity`,
all specifying different "types" of balances. The total balance of the
account needs to be derived from these.

This commit introduces `enum AccountBalanceCode` which is replicated
from the Ukob documentation. The values that are useful to us, based on
the description probided på Ukob, are then placed in an ordered list
with the most useful item first. Each AccountBalanceResponse will then
use this list to give the most useful account balance when `getBalance()`
is invoked.