fea(ukob-ais) Added fallback account name strategy
If `AccountEntity.nickName` is not set the account will now fall back
on `AccountIdentifierEntity.name`. This behaviour has is consistent with
 the Ukob documentation where it specifies that:

- `AccountEntity.nickName` is the name assigned by the account owner.
- `AccountIdentifierEntity.name` is the name given to the account by
agreement between the bank and owner.

This commit also fixes a potential problem with how the preferred
`AccountIdentifierEntity` is retrieved. Since `Stream` makes no strong
guarantees on the order of operations the solution provided in this
commit should be safer.