fix(ukob) Pagination url can now be complete or partial
Previously it was assumed that a pagination link would only contain a
path and query, as can be seen in the Ukob modelo example. But Lloyds
gives us a complete url. It is likely some banks will give use
complete urls and some will give us a path and query, so our code needs
to support both.

This commit add logic that checks if the given pagination link is a
partial url, in which case it concats it onto the apiBase. If the link
is a complete url it will use that directly instead.