fix(Norwegian): Parse recent payments that are not yet billed
We previously filtered out all transactions from the pending list that are booked since they are also present in the next transaction fetch (billed transactions for the current month). However, this also filtered out recent payments, they are booked but have not yet been added to the billed transactions list. Added a condition that looks at the accountTransactionId which should be 0, and that the transactionTypeText is "betalning".