fix(danskebank) change endpoint used in session handler
We have since the last couple of days seen errors due to the session handler. In the network traffic we can see the request made in the session handler looks to be OK,
but the next request (to account listing) fails with the same error we get when a user is not logged in (it's a generic error saying the service is unavailable at the moment).

Our theory is that danskebank has changed something making the endpoint we previously used (/extend) is not reliable when it comes to determining if the user is logged in or not.