fix(http/signature) sign the hash of headers and refactorings
Changed the code that generates a request signature to NOT include all raw headers but instead hash the headers (the same way body is hashed).
Without this fix the headers grows to more than 2 times the size when we attach our signature (headers x2 + signature).

Also refactored the code a bit and changed how the body is hashed:
instead of converting the input stream to a String (UTF8) and then back to a byte array I changed to to interpret it as a byte array in the first place.