fix(Refresh) Make use of cache instead of deprecated update method
Before we used `updateTransactions` method on `baseContext`, but this
method was deprecated. So the use of old method was removed and now
method `cacheTransactions` is used.
However this created a problem that creating separate method for
`LoanAccount` was supposed to solve. Because of use of generic
`Account`, when creating `AccountFeatures` the detailed type of the
class was lost, proper features for loans were not set.
To solve that I created a static method that replaces 3 overloaded
methods and decides on `AccountFeature` creation based of detailed type
of the account passed to it.