fix(AgentTests): Remove assertion for non empty portfolios
From what I investigated, there are cases when someone owns an
investment account, but does not have any portfolios - new account,
everything sold. Currently existing code broked the testing when an
ambassador had such a account. I did not find any other places that
needed to be change for it to work correctly.