fix(ukob/date-parsing) Reverted to using java.util.Date
ZonedDateTime has a bug that causes date parsing to crash on
short time offset, eg. `+01` instead of `+01:00` or `+1000` even though
this is valid according to documentation and ISO8601 standard.

For the time being we reverted back to java.util.Date which can handle
the format, even though we loose time and time offset info. But this
is not data we process right now anyway.

The bug mentioned above has apparently been fixed in jdk 9.
https://bugs.openjdk.java.net/browse/JDK-8032051