fix(FI/Danske): Skip xappculture header to satisfy header limit
Danske bank in Finland has a header limit of 17, when the different markets had different configs we removed the x-app-culture header for Finland. Now with the common filter we add it again. This commit adds a check for is the x-app-culture header should be added, in the Finland config that check will return false.