fea(currentModule) remove new modules
these modules are introduced for multi client support, they are not used and should not be used
remove introducing these module so we won't accidentally break the current aggregation. e.g. we might accidentally set it locally but data is not in production.
we can accidentally introduce changes that cause production to have NPE