Added ErrorResponse fo incorrect user credentials response
There were no 404 HTTP error when input wrong username/password so the response is now saved as a string and then serialised to json. Exception is thrown if the string response does not contain a ticket. And if the error code number is not 1025 then the number and error description is logged.