refactor(LockAgentWorkerCommand): Lock on credentialsId instead of provider
There is no longer need to lock over the whole provider as we have another lock specifically for bankID that locks the user. This way we don't get problems with creates of a new credential for the same provider if there is another credentials that holds the lock. This was an issue for ukob.