fix(LockAgentWorkerCmd): Decrease wait time to acquire lock for UK
We've discovered issues for UK with two refresh requests coming in at the same time. The first request successfully executes and updates the credential, but once the second request manages to acquire the lock and start the refresh it works on stale data and starts the manual refresh again. This do-over puts the credential in AUTHENTICATION_ERROR which we don't want. This fix is ugly and hacky and should be replaced with a more robust solution.