fix(ukob/balance): switched preferred balance type
Switched to prefer `INTERIM_BOOKED` over `INTERIM_AVAILABLE` since this
seems to be the value expected by users.

`INTERIM_BOOKED` and `INTERIM_AVAILABLE` seems to be the most common
account balance types we get from the banks so far.

The different types are explained here:
https://openbanking.atlassian.net/wiki/spaces/DZ/pages/128909480/Balances+v2.0.0

But the only difference in the description of the types are the words
"available" and "booked".

Looking further into the concept of "booked" and "available" it seems
like this is the case:

booked = accountBalance - pendingTransactions + incomingDeposits

available = accountBalance