fix(uk/revolut): removed failed transactions
Transactions with state == FAILED were previously included. This commit
introduces a check that filter these transaction.

It also refactors some of the code in RevolutTransactionFetcher,
but except for the added state check there should be no behaviour change.

The code could probably be refacored a lot more, but I don't want to
risk introducing a behaviour change at this time.