fix(ukob/balance-sign): corrected account balance sign
The credit/debit indicator indicates the sign of the balance rather than
if the balance has credit, as we previously assumed.

This commit sets the sign of the balance based on credit/debit indicator.

It also addresses an issue where some banks will also include the sign
in the value itself while others will only rely on the credit/debit
indicator. We now ignore any sign included in the value and apply the
sign using the credit/debit indicator.