fix(stale-credential): Temporary fix for stale credential data
It is possible that several refresh requests for the same credential to
exists on the queue at the same time, if the credentials are updated
(such as usage of refresh token) the next request will have stale data.

This was a known issue before and a cache was introduced as a
workaround. Caching the credentials with a high enough CACHE_EXPIRE_TIME
allows subsequent request to exchange its stale data for the updated
data in the cache. The logic reading from the cache can be found in
`CredentialsCrypto.java`.

This commit increases the expire time of the cache from 10 min to
20 min, since 10 min still allowed the cache to expire before subsequent
queued requests to arrive at an empty cache and therefor end up with
stale data.

This is not a permanent fix and should be resolved by overhauling the
credential read/write logic such that a request will never have stale
data in the first place.