fix(norwegian): parse recent payments
"Recent payments" were previously filtered out from pending transactions which resulted in them not being aggregated until they existed in the billed transactions list days later.

A previous fix (see #942) did parse recent payments from pending transactions, however, once it was listed in the billed transactions it was read again resulting in duplicate transactions. So the fix got reverted.

This fix re-uses the previous fix but it also saves the external Id of all added transactions and checks when it reads from the billed transactions that it filters out the transactions that already has an existing external Id in the list to avoid duplicates.