fix(storage): Added string type checking to Storage class
If Storage::get(String, Class<T>) is called and
the second argument is String.class, the method
would attempt to deserialize the stored object
(which in this case is a string), even though it
was not serialized when it was put.
Deserialization would fail and the method would
return Optional.empty().