fix(revolut): Removed check for wallet ref
It seems like it was previously assumed that in order to build a valid
account from the Revolut `Wallet, Topup, Pocket` -format a Topup needed
to share a reference number with the Wallet.

From looking at logs and user feedback this does not seem to be
the case. The result was that we filtered out Topups that should be
valid, in turn resulting in accounts with 0 balance even though the
user did in fact have money on that account.

This commit removes the above mentioned reference check and removes a
write to temporary storage, since this was never read from.