fix(revolut/bic): Added bic as a possible identifier
Previously the account entity assumed that `iban` or `accountNumber`
would be set.

It turns out that `iban` or `accountNumber` or `bic` can be set, this
caused account builder to fail since it was supplied with
`accountNumber == null`.

This commit adds `bic` as a possible identifier and refactors the
getter to be clearer and throw a more obvious exception if this
value is still `null` after this fix.