fix(BEC loans): Handle no details exist error
We get "No details exist." as message when trying to fetch loans (apiClient.fetchLoans()) for users that, I assume, don't have any loans. Added handling of this error so that we return an empty list if we get this message as well. Also added logging for both messages that we look at, so that we can see how often we come across each.