fix(ES/Santander): Fix XML serialization to string
With the changes to session storage (12/12) we don't deserialize from String if the value is fetched as a String. The lead to us getting a "String in a String" when fetching the login response from the session storage. In that format we could not convert the XML String to JSON with Jackson.