fix(revolut/pocket-structure) Reworked how accounts are constructed
**Summary**

This commit changes how accounts are constructed from revolut pockets,
this fixes an issue where the account number type would depend on the
order we recieved accounts in.

It also tries to make the flow a bit easier to follow and adds
`holderName` to the accounts.

**Detailed Description**

Revolut has a structure where an account is split up into a
storage part (pocket) and a recieving part (account). The pocket holds
the funds while the account contains an account identifier.

* Only one pocket per currency may exist.

* For every pocket there is at least one account which specifies IBAN.

* If a pocket can recieve transfers by several different identifiers
types (IBAN, BIC, SC/AN) there will be one account per identifier type.

In order to convert this structure to a "Tink account" the
`RevolutTransactionalAccountFetcher` will first map every currency to
an IBAN by looking at all accounts. Then it will match each pocket to
an IBAN by currency which is enough to construct a "Tink account".