fix(nordea/investment): Liquidity accounts merged with investment
Liquidity accounts were previously converted into checking/savings
accounts. After complaints from customers it was decided that
the luquidity accounts would instead represent the `cashValue` of the
corresponding investment.

This was solved by checking if each account belonged to an investment,
if this is the case we store the balance of that account in a map
intead of constructing a tink account. This map is then queried when
investments are constructed.