build(ci): Remove --batch bazel flag from ci scripts
The --batch flag is deprecated, from the bazel documentation: "If set,
Bazel will be run as just a client process without a server, instead of
in the standard client/server mode. This is deprecated and will be
removed, please prefer shutting down the server explicitly if you wish
to avoid lingering servers."
Running in this mode is undesirable anyways because it does not allow us
to reuse cache.