fea(agent/utils) vasco/digipass utility
This util can be used to calculate OTPs.

It currently only supports time based OTPs, not counter. A future improvement will be to merge it with KBC's digipass implementation (they differ due to configuration in the static vector).

The Digipass class is designed similar to the vasco digipass SDK. Use serialize and deserialize to save/load the "device" to/from persistent storage.
I chose to have a specific serialization/deserialization method instead of serializing the whole object in order to make it future proof.