fix(Bankia): Correct logic for when to stop transaction fetch
We got an exception when toDate was 2017-04-22 and we looked at if it was after 2018-01-29, and since it wasn't we threw an exception. Changed to check if toDate is before the one year ago date. Then we throw an exception if the toDate is within 12 months, and stop fetching without error if it's not.