fix(danskebank) eternal pagination loop
This is a quick fix that breaks out of a eternal loop that can sometimes
happen with danskebank pagination.

If we go outside of the allowed pagination date range danskebank will
just return the last transactions over and over again.

This commit breaks after paginating 2 years, which seems to be what can
be seen in the app.