fix(ES ING): Fix transaction pagination offset
We got an error about invalid offset when paginating for a beta user. Now when we fetch for a month; if the amount of returned transactions is less than the total amount of transactions we fetch again for this month, with the offset set to current offset + the number of transactions returned. A case would be that for a first round the offset is 0, 25 transactions are returned but the total is 53. We then fetch again, with the offset = 0 + 25 = 25. The expected response is 25 transactions, with total of 53. So we fetch again, with offset = 25 + 25 = 50. Then we should get the last 3 transactions. Now currOffset, which is 50 + 3 = 53, isn't less than the total of 53, so the while loop ends.